Fix height value written in the size register
Apparently, it has to be (height + 1). For instance, if the output
buffer is 320x240, then a height of 241 should be written. Otherwise,
the IPU will skip the last line. This is particularly visible in
direct-connect mode, as the frame moves from bottom to top at a speed
of 60 pixels per second.